Extract DocumentHistory and move storage-metadata tests to their own file
