+++ editor/src/messages/portfolio/document/mod.rs
@@ -1,5 +1,8 @@
+mod document_history;
 mod document_message;
 mod document_message_handler;
+#[cfg(test)]
+mod storage_metadata_tests;
 #[cfg(test)]
 mod storage_round_trip_tests;
 
@@ -12,6 +15,7 @@ pub mod properties_panel;
 pub mod resource;
 pub mod utility_types;
 
+pub(crate) use document_history::DocumentHistory;
 #[doc(inline)]
 pub use document_message::{DocumentMessage, DocumentMessageDiscriminant};
 pub(crate) use document_message_handler::diff_networks;
